RaspberryPi-Tub-Pipeline/requirements.txt
fastapi

fastapi
uvicorn[standard]
kafka-python
influxdb-client
cassandra-driver
pytest
schedule
python-dateutil
